#S07P31B102-223 feat. 거래 구현중

+++ Back/selly-trade-service/src/main/java/com/nct/sellytradeservice/domain/dto/NftPieceDto.java
@@ -6,6 +6,6 @@ import lombok.Getter;
 public class NftPieceDto {
   private Long articleId;
   private Long userId;
-  private Long nftPieceCnt;
+  private Integer nftPieceCnt;
   private double avgPrice;
 }
